+++ CHANGELOG.md
@@ -24,9 +24,14 @@
 * Properly serialize array settings when running `config-dump`.  
   [Samuel Giddins](https://github.com/segiddins)
 
-* Add iMessage extensions. 
+* Add iMessage extensions.  
   [wade0317](https://github.com/wade0317)
   [#723](https://github.com/CocoaPods/Xcodeproj/pull/723)
+
+* Fix errors when using mutually recursive build settings.  
+  [revolter](https://github.com/revolter)
+  [#727](https://github.com/CocoaPods/Xcodeproj/pull/727)
+
 
 ## 1.13.0 (2019-10-16)